Refactor auth and login flow for smoother user experience
This commit refactors our authentication handling and login flow to enhance the user experience. The idToken property was removed in the oauth configuration in the auth API since it was not needed. Where the user ID was previously obtained from the profile sub, it's now from profile email. Moreover, consistency was improved through formatting changes and unwanted console logs were removed.

In the homepage component, the signIn and signOut methods were updated for a smoother user experience. It now uses a different method of checking and handling the session status, providing a better way to redirect users who are not authenticated, offering a more structured and user-friendly approach.