Commit message:
Fix signIn function and adjust import syntax

In this commit, the signIn function has been fixed to use the 'pres' parameter which could be necessary due to the system environment setup, as previous settings led to unexpected behavior. Moreover, the import syntax has been adjusted in order to follow the code conventions in the rest of the application. All spacing in import statements has been removed to follow the project's style guide. The changes to pnpm